requirements.txt
streamlit
semantic-kernel
python-dotenv
azure-core
azure-search-documents